+++ src/stitching/introspectSchema.ts
@@ -1,15 +1,23 @@
-import { GraphQLSchema } from 'graphql';
+import { GraphQLSchema, parse } from 'graphql';
 import { introspectionQuery, buildClientSchema } from 'graphql';
-import { Fetcher } from './makeRemoteExecutableSchema';
+import { ApolloLink, execute, makePromise } from 'apollo-link';
+import { Fetcher, fetcherToLink } from './makeRemoteExecutableSchema';
+
+const parsedIntrospectionQuery = parse(introspectionQuery);
 
 export default async function introspectSchema(
-  fetcher: Fetcher,
+  link: ApolloLink | Fetcher,
   context?: { [key: string]: any },
 ): Promise<GraphQLSchema> {
-  const introspectionResult = await fetcher({
-    query: introspectionQuery,
-    context,
-  });
+  if (!(link as ApolloLink).request) {
+    link = fetcherToLink(link as Fetcher);
+  }
+  const introspectionResult = await makePromise(
+    execute(link as ApolloLink, {
+      query: parsedIntrospectionQuery,
+      context,
+    }),
+  );
   if (introspectionResult.errors || !introspectionResult.data.__schema) {
     throw introspectionResult.errors;
   } else {